Adding data changes.

+++ scripts/data_manipulation/data_v2_format_check.py
@@ -2,6 +2,7 @@ import argparse
 import json
 import numbers
 import glob
+import os
 
 import logging
 logger = logging.getLogger(__name__)
@@ -138,7 +139,7 @@ def fix_keys_to_doc_name(objs,f):
 
     for obj in objs:
         if obj['pk'] != "{}_{}".format(slugify(f['doc']),slugify(obj['fields']['name'])):
-            if f['filename']=='Spell.json':
+            if f['filename']=='Item.json':
                 logger.warning("{} changing to doc_name format".format(f['path']))
                 pk_value = "{}_{}".format(obj['fields']['document'],slugify(obj['fields']['name']))
                 logger.warning("CHANGING PK TO {}".format(pk_value))
@@ -149,15 +150,15 @@ def fix_keys_to_doc_name(objs,f):
 
 
         related_path = "{}/{}/{}/{}/{}/".format(f['root'],f['dir'],f['schema'],f['publisher'],f['doc'])
-        related_filenames = ['SpellCastingOption.json']
+        related_filenames = ['ItemSet.json']
 
     for obj in objs_fixed:
         for related_file in related_filenames:
             logger.warning("CHANGING RELATED PK IN {} TO {}".format(related_file,obj['pk']))
-            refactor_relations(related_path+related_file,"parent",obj['former_pk'], obj['pk'])
+            refactor_relations(related_path+related_file,"items",obj['former_pk'], obj['pk'])
         obj.pop('former_pk')
 
-    if f['filename']=='Spell.json':    
+    if f['filename']=='Item.json':    
         with open(f['path'],'w',encoding='utf-8') as wf:
             json.dump(objs_fixed,wf,ensure_ascii=False,indent=2)
         pass
@@ -188,15 +189,23 @@ def fix_keys_to_parent_level(objs,f):
 
 def refactor_relations(filename, key, former_pk, new_pk):
     refactored_objects = []
-    with open(filename, 'r', encoding='utf-8') as f:
-        objs = json.load(f)
-        for obj in objs:
-            if obj['fields'][key] == former_pk:
-                obj['fields'][key] = new_pk
-            refactored_objects.append(obj)
+    if os.path.isfile(filename):
+        with open(filename, 'r', encoding='utf-8') as f:
+            objs = json.load(f)
+            if key == "parent":
+                for obj in objs:
+                    if obj['fields'][key] == former_pk:
+                        obj['fields'][key] = new_pk
+                    refactored_objects.append(obj)
+            if key == "items":
+                for obj in objs:
+                    if former_pk in obj['fields'][key]:
+                        obj['fields'][key].remove(former_pk)
+                        obj['fields'][key].append(new_pk)
+                    refactored_objects.append(obj)
 
-    with open(filename,'w', encoding='utf-8') as o:
-        json.dump(refactored_objects,o, ensure_ascii=False,indent=2)
+        with open(filename,'w', encoding='utf-8') as o:
+            json.dump(refactored_objects,o, ensure_ascii=False,indent=2)
 
 def check_keys_doc_parent_name(objs,f):
     for obj in objs:
